refactor(frontend/git): extract Git components into separate modules
- Split GitPane into smaller reusable components
- Create GitCommitSection for commit functionality
- Create GitRepositoryStatus for repository state handling
- Create GitStagedChanges and GitUnstagedChanges components
- Create GitStatusItem for file status display
- Improve code organization and maintainability
